updates for facades files for updated schema testing
